separate schema from datasource
